move files to webpack folder
