+++ lib/webpack/create-webpack-config.js
@@ -8,8 +8,8 @@ const resolvePkg = require('resolve-pkg');
 const urlJoin = require('url-join');
 const fs = require('fs');
 
-const safeWebpackUglifier = require('./utils/safe-webpack-uglifier');
-const { WEBPACK_ASSETS_BASENAME } = require('./constants');
+const safeWebpackUglifier = require('../utils/safe-webpack-uglifier');
+const { WEBPACK_ASSETS_BASENAME } = require('../constants');
 
 function createWebpackConfig(urc) {
   const jsEntryDir = path.dirname(urc.jsEntry);
@@ -110,7 +110,7 @@ function createBabelLoader(urc) {
   const loader = {
     loader: 'babel-loader',
     options: {
-      presets: [require.resolve('../packages/babel-preset-mapbox')],
+      presets: [require.resolve('../../packages/babel-preset-mapbox')],
       cacheDirectory: true,
       babelrc: false,
       compact: false
