+++ lib/webpack/write-webpack-stats.js
